+++ toolchains/common_toolchains/release_toolchain.cmake
@@ -5,7 +5,7 @@ include(${CMAKE_CURRENT_LIST_DIR}/importable_defaults.cmake)
 set(CMAKE_BUILD_TYPE Release)
 
 # Compiler options
-set(CMAKE_CXX_FLAGS_INIT "-Wall -Werror=unused-local-typedefs -Werror=unused-variable -Wno-sign-compare -ffast-math")
+set(CMAKE_CXX_FLAGS_INIT "-Wall -Werror=unused-local-typedefs -Werror=unused-variable -Wno-sign-compare -fno-signed-zeros -ffinite-math-only")
 
 # Activate/deactivate parts of the code
 set(GYSELALIBXX_ACTIVATE_RESTART_TESTS OFF)