+++ tests/geometryXVx/spatialadvection.cpp
@@ -5,7 +5,6 @@
 #include "bsl_advection_x.hpp"
 #include "geometry_xvx.hpp"
 #include "spline_definitions_xvx.hpp"
-#include "spline_interpolator.hpp"
 
 
 
@@ -101,9 +100,8 @@ TEST(SpatialAdvection, SplineBatched)
     ddc::PeriodicExtrapolationRule<X> bv_x_min;
     ddc::PeriodicExtrapolationRule<X> bv_x_max;
     SplineXEvaluator const spline_x_evaluator(bv_x_min, bv_x_max);
-    PreallocatableSplineInterpolator const
-            spline_x_interpolator(builder_x, spline_x_evaluator, meshXVx);
-    BslAdvectionSpatial<GeometryXVx, GridX> const spline_advection_x(spline_x_interpolator);
+    BslAdvectionSpatial<GeometryXVx, GridX, SplineXBuilder, SplineXEvaluator> const
+            spline_advection_x(builder_x, spline_x_evaluator);
     double const err
             = SpatialAdvection<GeometryXVx, GridX>(spline_advection_x, idx_range_x, idx_range_vx);
     EXPECT_LE(err, 1.e-6);