+++ src/advection/bsl_advection_vx.hpp
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: MIT
 #pragma once
+#include <type_traits>
+
 #include <ddc/ddc.hpp>
 
 #include "ddc_alias_inline_functions.hpp"
@@ -17,9 +19,12 @@ template <
         class Geometry,
         class GridV,
         concepts::InterpolationBuilder FunctionBuilder,
-        concepts::InterpolationEvaluator FunctionEvaluator>
-class BslAdvectionVelocity : public IAdvectionVelocity<Geometry, GridV>
+        concepts::InterpolationEvaluator FunctionEvaluator,
+        class DataType = double>
+class BslAdvectionVelocity : public IAdvectionVelocity<Geometry, GridV, DataType>
 {
+    static_assert(std::is_floating_point_v<DataType>);
+
     using IdxRangeFdistribu = typename Geometry::IdxRangeFdistribu;
     using IdxRangeSpatial = typename Geometry::IdxRangeSpatial;
     using IdxSpatial = typename IdxRangeSpatial::discrete_element_type;
@@ -62,10 +67,10 @@ public:
      * @param[in] dt Time step
      * @return A reference to the allfdistribu array containing the value of the function at the coordinates.
      */
-    Field<double, IdxRangeFdistribu> operator()(
-            Field<double, IdxRangeFdistribu> const allfdistribu,
-            Field<const double, IdxRangeSpatial> const electric_field,
-            double const dt) const override
+    Field<DataType, IdxRangeFdistribu> operator()(
+            Field<DataType, IdxRangeFdistribu> const allfdistribu,
+            ConstField<DataType, IdxRangeSpatial> const electric_field,
+            DataType const dt) const override
     {
         using IdxRangeBatch = ddc::remove_dims_of_t<IdxRangeFdistribu, Species, GridV>;
         using IdxBatch = typename IdxRangeBatch::discrete_element_type;
@@ -78,9 +83,9 @@ public:
 
         IdxRangeSpaceVelocity batched_feet_idx_range(idx_range);
 
-        FieldMem<double, IdxRangeFunctionDeriv> derivs_min(
+        FieldMem<DataType, IdxRangeFunctionDeriv> derivs_min(
                 m_function_builder.batched_derivs_xmin_domain(batched_feet_idx_range));
-        FieldMem<double, IdxRangeFunctionDeriv> derivs_max(
+        FieldMem<DataType, IdxRangeFunctionDeriv> derivs_max(
                 m_function_builder.batched_derivs_xmax_domain(batched_feet_idx_range));
         ddc::parallel_fill(derivs_min, 0.);
         ddc::parallel_fill(derivs_max, 0.);
@@ -94,16 +99,16 @@ public:
         IdxRangeBatch batch_idx_range(idx_range);
 
         ddc::host_for_each(idx_range_sp, [&](IdxSp const isp) {
-            double const charge_proxy
+            DataType const charge_proxy
                     = charge(isp); // TODO: consider proper way to access charge from device
-            double const sqrt_me_on_mspecies = std::sqrt(mass(ielec()) / mass(isp));
+            DataType const sqrt_me_on_mspecies = std::sqrt(mass(ielec()) / mass(isp));
             ddc::parallel_for_each(
                     Kokkos::DefaultExecutionSpace(),
                     batch_idx_range,
                     KOKKOS_LAMBDA(IdxBatch const ib) {
                         IdxSpatial const ix(ib);
                         // compute the displacement
-                        double const dvx
+                        DataType const dvx
                                 = charge_proxy * sqrt_me_on_mspecies * dt * electric_field(ix);
 
                         // compute the coordinates of the feet