+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `-ffast-math` flag to toolchains.
+- Add `-fno-signed-zeros -ffinite-math-only` flags to toolchains.
 
 ### Fixed