+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `-fno-signed-zeros -ffinite-math-only` flags to toolchains.
+
 ### Fixed
 
 ### Changed