+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use generic binaries in the CPU Spack toolchain.
 - Remove Kokkos-related submodules.
 - Remove recursion from submodules synchronisation.
+- Add missing `module purge` in Persee toolchains.
 
 ### Deprecated